ml-model-api/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
werkzeug==3.0.1
bleach==6.1.0
redis==5.0.1
Pillow==10.2.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.8.1.78
